Rename editor tests namespace from "Editor" to "Editing"
