+++ osu.Game.Tests/Visual/Editing/TestSceneBeatDivisorControl.cs
@@ -14,7 +14,7 @@ using osu.Game.Screens.Edit.Compose.Components;
 using osuTK;
 using osuTK.Input;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     public class TestSceneBeatDivisorControl : OsuManualInputManagerTestScene
     {
