+++ osu.Game.Tests/Visual/Editing/TestSceneDistanceSnapGrid.cs
@@ -13,7 +13,7 @@ using osu.Game.Screens.Edit.Compose.Components;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     public class TestSceneDistanceSnapGrid : EditorClockTestScene
     {
