+++ osu.Game.Tests/Visual/Editing/TestSceneEditorComposeRadioButtons.cs
@@ -7,7 +7,7 @@ using NUnit.Framework;
 using osu.Framework.Graphics;
 using osu.Game.Screens.Edit.Components.RadioButtons;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     [TestFixture]
     public class TestSceneEditorComposeRadioButtons : OsuTestScene
