+++ osu.Game.Tests/Visual/Editing/TestSceneEditorSummaryTimeline.cs
@@ -10,7 +10,7 @@ using osu.Game.Rulesets.Osu;
 using osu.Game.Screens.Edit.Components.Timelines.Summary;
 using osuTK;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     [TestFixture]
     public class TestSceneEditorSummaryTimeline : EditorClockTestScene
