+++ osu.Game.Tests/Visual/Editing/TestSceneHitObjectComposer.cs
@@ -20,7 +20,7 @@ using osu.Game.Screens.Edit;
 using osu.Game.Screens.Edit.Compose.Components;
 using osuTK;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     [TestFixture]
     public class TestSceneHitObjectComposer : EditorClockTestScene
