+++ osu.Game.Tests/Visual/Editing/TestScenePlaybackControl.cs
@@ -9,7 +9,7 @@ using osu.Game.Beatmaps;
 using osu.Game.Screens.Edit.Components;
 using osuTK;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     [TestFixture]
     public class TestScenePlaybackControl : OsuTestScene
