+++ osu.Game.Tests/Visual/Editing/TestSceneTimelineBlueprintContainer.cs
@@ -7,7 +7,7 @@ using NUnit.Framework;
 using osu.Framework.Graphics;
 using osu.Game.Screens.Edit.Compose.Components.Timeline;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     [TestFixture]
     public class TestSceneTimelineBlueprintContainer : TimelineTestScene
