+++ osu.Game.Tests/Visual/Editing/TestSceneTimelineTickDisplay.cs
@@ -8,7 +8,7 @@ using osu.Game.Screens.Edit.Compose.Components;
 using osu.Game.Screens.Edit.Compose.Components.Timeline;
 using osuTK;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     [TestFixture]
     public class TestSceneTimelineTickDisplay : TimelineTestScene
