+++ osu.Game.Tests/Visual/Editing/TestSceneZoomableScrollContainer.cs
@@ -15,7 +15,7 @@ using osu.Game.Screens.Edit.Compose.Components.Timeline;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Tests.Visual.Editor
+namespace osu.Game.Tests.Visual.Editing
 {
     public class TestSceneZoomableScrollContainer : OsuManualInputManagerTestScene
     {
